SEPARATE INS and DEL settings and other
Instead of randomly deciding the number of insertions and deletions
from the INDEL_ parameters, separate the settings so that it is
possible to choose the desired number of insertions and deletions

Add dup_num info field in header, to avoid errors when using
bcftools. (Dummy description string).

Add gitignore

Other minor modifications